ref(sidebar): Move sidebar outside-click close logic to SidebarPanel
